requirements.txt
flask
gunicorn
markdown
markupsafe
psycopg2-binary
python-dotenv
